+++ src/library/CMakeLists.txt
@@ -21,8 +21,9 @@ set( clFFT.Source	transform.cpp
 								repo.cpp
 								generator.stockham.cpp
 								generator.transpose.gcn.cpp
-								generator.transpose.nonsquare.cpp
-								generator.transpose.square.cpp
+								generator.transpose.cpp
+								action.transpose.nonsquare.cpp
+								action.transpose.square.cpp
 								generator.copy.cpp
 								lifetime.cpp
 								fft_binary_lookup.cpp
@@ -44,8 +45,9 @@ set( clFFT.Headers	private.h
 					generator.h
 					generator.stockham.h
 					generator.transpose.gcn.h
-					generator.transpose.nonsquare.h
-					generator.transpose.square.h
+					generator.transpose.h
+					action.transpose.nonsquare.h
+					action.transpose.square.h
 					fft_binary_lookup.h
 					md5sum.h
 					../include/stdafx.h
