+++ src/library/action.transpose.nonsquare.h
@@ -21,7 +21,6 @@
 #include "repo.h"
 #include "plan.h"
 
-#define AVAIL_MEM_SIZE 32768 
 #endif
 
 #pragma once
